requirements.txt
acme==1.1.0
appdirs==1.4.4
apturl==0.5.2
astroid==2.4.2
attrs==19.3.0
backcall==0.2.0
bcrypt==3.1.7
bleach==3.1.5
blinker==1.4
Brlapi==0.7.0
certbot==0.40.0
certbot-nginx==0.40.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==8.1.3
colorama==0.4.3
command-not-found==0.3
ConfigArgParse==0.13.0
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==2.2.3
Flask-Login==0.6.2
Flask-Principal==0.4.0
flask-sock==0.4.0
Flask-WTF==1.1.1
future==0.18.2
greenlet==2.0.2
gunicorn==20.0.4
h11==0.12.0
httplib2==0.14.0
idna==2.8
importlib-metadata==6.6.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==2.1.2
jedi==0.17.2
Jinja2==3.1.2
josepy==1.2.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
kazam==1.5.3
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.2
matplotlib==3.1.2
mccabe==0.6.1
meld==3.20.2
mistune==0.8.4
mock==3.0.5
monotonic==1.5
more-itertools==4.2.0
mysql-connector==2.2.9
mysql-connector-python==8.0.21
nbconvert==5.6.1
nbformat==5.0.7
netifaces==0.10.4
notebook==6.0.3
numpy==1.23.1
oauthlib==3.1.0
olefile==0.46
packaging==20.4
pandocfilters==1.4.2
paramiko==2.6.0
parsedatetime==2.4
parso==0.7.1
pbr==5.4.5
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
Pivy==0.6.5
plotly==4.4.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.6.1
ptyprocess==0.6.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.6.1
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==1.7.1
pylint==2.5.3
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.16.0
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-xlib==0.23
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.1
reportlab==3.5.34
requests==2.22.0
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
retrying==1.3.3
SecretStorage==2.3.1
Send2Trash==1.5.0
simple-websocket==0.4.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
SQLAlchemy==2.0.10
ssh-import-id==5.10
supervisor==4.2.2
systemd-python==234
terminado==0.8.3
testpath==0.4.4
testresources==2.0.1
tk==0.1.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typing-extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.28
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.12.1
wsproto==1.0.0
WTForms==2.3.3
wxPython==4.0.7
xkit==0.0.0
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.0
zope.interface==4.7.1
